Stitch together segments of time series measurements

Parses a Tecan Infinite 200 Pro plate reader Excel file that contains segments of a time series spread over multiple sheets and automatically stitches them together accounting for the time gap between them

Usage

tecan_unite(xlsx_file, segments = NULL, skip = 0)

Arguments

xlsx_file

(character) path to Excel file

segments

(numeric) number of segments to stitch together, starting from the last sheet (default: number of sheets)

skip

(numeric) number of sheets to skip, starting with the last sheet (default: 0). Note that empty sheets will be skipped automatically

Value

A tibble::tibble() containing tidy data

Examples

tecan_unite(
  system.file(
    "extdata",
    "tecan_time_series_segments.xlsx",
    package = "tecanr"
  )
)
#> ℹ Time series detected
#> ℹ Multiple reads per well detected
#> ℹ Time series detected
#> ℹ Multiple reads per well detected
#> # A tibble: 3,432 × 9
#>    well   time  temp  mean st_dev  x1_2  x0_1  x1_0  x2_1
#>    <chr> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 B1        0  30.6 0.131 0.0254 0.131 0.111 0.114 0.166
#>  2 B1      900  30.4 0.129 0.0216 0.132 0.112 0.115 0.159
#>  3 B1     1500  30.5 0.133 0.0226 0.146 0.113 0.114 0.157
#>  4 B1     2100  30.7 0.132 0.0226 0.143 0.113 0.115 0.160
#>  5 B1     2700  30.6 0.133 0.0220 0.146 0.114 0.116 0.158
#>  6 B1     3300  30.2 0.133 0.0208 0.144 0.114 0.116 0.156
#>  7 B1     3900  30.6 0.135 0.0227 0.147 0.115 0.117 0.161
#>  8 B1     4500  30.6 0.134 0.0210 0.146 0.116 0.118 0.158
#>  9 B1     5100  30.5 0.135 0.0208 0.147 0.117 0.119 0.159
#> 10 B1     5700  30.5 0.136 0.0201 0.147 0.118 0.120 0.158
#> # ℹ 3,422 more rows